+++ packages/gatsby-plugin-treat/gatsby-node.js
@@ -1,63 +1,35 @@
 const TreatPlugin = require('treat/webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
-exports.onCreateBabelConfig = ({stage, actions}) => {
+exports.onCreateBabelConfig = ({ actions }) => {
   actions.setBabelPlugin({
-    name: `babel-plugin-treat`,
+    name: 'babel-plugin-treat',
   });
-}
+};
 
-exports.onCreateWebpackConfig = ({stage, actions}) => {
-  let config = {};
+exports.onCreateWebpackConfig = (
+  { stage, loaders, actions },
+  { plugins, ...pluginOptions },
+) => {
+  if (stage === 'develop-html') return;
 
-  const identNameConfig =
-    stage === `develop` || stage === `develop-html`
-      ? {
-          localIdentName: '[name]-[local]_[hash:base64:5]',
-          themeIdentName: '_[name]-[local]_[hash:base64:4]',
-        }
-      : {
-          localIdentName: '[hash:base64:5]',
-          themeIdentName: '[hash:base64:4]',
-        };
-
-  switch (stage) {
-    case `develop`:
-      config = {
-        plugins: [
-          new TreatPlugin({
-            ...identNameConfig
-          }),
-        ]
+  const defaultPluginOptions = stage.includes('develop')
+    ? {
+        localIdentName: '[name]-[local]_[hash:base64:5]',
+        themeIdentName: '_[name]-[local]_[hash:base64:4]',
+      }
+    : {
+        localIdentName: '[hash:base64:5]',
+        themeIdentName: '[hash:base64:4]',
       };
-      break;
 
-    case `develop-html`:
-      break;  
-    
-    case `build-javascript`:
-      config = {
-        plugins: [
-          new TreatPlugin({
-            ...identNameConfig,
-            outputLoaders: [MiniCssExtractPlugin.loader]
-          })
-        ]
-      };
-      break;
-
-    case `build-html`:
-      config = {
-        plugins: [
-          new TreatPlugin({
-            ...identNameConfig,
-            outputCSS: false
-          }),
-        ]
-      };
-      break;
-
-  }
-
-  actions.setWebpackConfig(config);
- }
+  actions.setWebpackConfig({
+    plugins: [
+      new TreatPlugin({
+        ...defaultPluginOptions,
+        ...pluginOptions,
+        outputCSS: !stage.includes('html'),
+        outputLoaders: [loaders.miniCssExtract()],
+      }),
+    ],
+  });
+};