+++ packages/gatsby-plugin-treat-example/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.1](https://github.com/seek-oss/treat/compare/v1.5.0...v1.5.1) (2020-07-27)
+
+**Note:** Version bump only for package gatsby-plugin-treat-example
+
+
+
+
+
+# [1.5.0](https://github.com/seek-oss/treat/compare/v1.4.3...v1.5.0) (2020-07-23)
+
+**Note:** Version bump only for package gatsby-plugin-treat-example
+
+
+
+
+
+## [1.4.3](https://github.com/seek-oss/treat/compare/v1.4.2...v1.4.3) (2020-07-13)
+
+**Note:** Version bump only for package gatsby-plugin-treat-example
+
+
+
+
+
+## [1.4.2](https://github.com/seek-oss/treat/compare/v1.4.1...v1.4.2) (2020-04-26)
+
+**Note:** Version bump only for package gatsby-plugin-treat-example
+
+
+
+
+
+## [1.4.1](https://github.com/seek-oss/treat/compare/v1.4.0...v1.4.1) (2020-04-26)
+
+**Note:** Version bump only for package gatsby-plugin-treat-example
+
+
+
+
+
+# [1.4.0](https://github.com/seek-oss/treat/compare/v1.3.0...v1.4.0) (2020-04-24)
+
+
+### Features
+
+* **gatsby-plugin:** Support webpack plugin options ([#100](https://github.com/seek-oss/treat/issues/100)) ([adc32f1](https://github.com/seek-oss/treat/commit/adc32f1))
+
+
+
+
+
 ## [1.2.4](https://github.com/seek-oss/treat/compare/v1.2.3...v1.2.4) (2020-01-31)
 
 **Note:** Version bump only for package gatsby-plugin-treat-example