+++ packages/gatsby-plugin-treat/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.2](https://github.com/seek-oss/treat/tree/master/packages/gatsby-plugin-treat/compare/v1.4.1...v1.4.2) (2020-04-26)
+
+
+### Bug Fixes
+
+* **gatsby-plugin:** Idiomatic SSR setup ([#108](https://github.com/seek-oss/treat/tree/master/packages/gatsby-plugin-treat/issues/108)) ([4dc6d9c](https://github.com/seek-oss/treat/tree/master/packages/gatsby-plugin-treat/commit/4dc6d9c))
+
+
+
+
+
+# [1.4.0](https://github.com/seek-oss/treat/tree/master/packages/gatsby-plugin-treat/compare/v1.3.0...v1.4.0) (2020-04-24)
+
+
+### Features
+
+* **gatsby-plugin:** Support webpack plugin options ([#100](https://github.com/seek-oss/treat/tree/master/packages/gatsby-plugin-treat/issues/100)) ([adc32f1](https://github.com/seek-oss/treat/tree/master/packages/gatsby-plugin-treat/commit/adc32f1))
+
+
+
+
+
 ## [1.2.3](https://github.com/seek-oss/treat/tree/master/packages/gatsby-plugin-treat/compare/v1.2.2...v1.2.3) (2020-01-15)
 
 **Note:** Version bump only for package gatsby-plugin-treat